Moved files following standard structure
