+++ src/assets/imgs/logo.png
