+++ src/components/Header.js
@@ -2,8 +2,8 @@ import React, { useState, useEffect } from 'react';
 import { Menu, X } from 'lucide-react';
 import ThemeToggle from './ThemeToggle';
 // import jujuLogo from '../assets/juju-logo.svg';
-import jujuLogo from '../assets/logo.svg';
-import './Header.css';
+import jujuLogo from '../assets/imgs/logo.svg';
+import '../assets/styles/Header.css';
 
 const Header = () => {
   const [isMenuOpen, setIsMenuOpen] = useState(false);
